split profile data
